+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/devices/jyou/BFH16DeviceCoordinator.java
@@ -16,7 +16,7 @@
     You should have received a copy of the GNU Affero General Public License
     along with this program.  If not, see <http://www.gnu.org/licenses/>. */
 
-package nodomain.freeyourgadget.gadgetbridge.devices.bfh16;
+package nodomain.freeyourgadget.gadgetbridge.devices.jyou;
 
 import android.app.Activity;
 import android.bluetooth.le.ScanFilter;
